Move overlay logic to class
